Changed Makefile to pass all macro definitions to lint. Changed Makefile to only include ncurses library if necessary. Changed code to obviate requirement of math library (for log10() function). Continued changes to use uint_t type, to allow both 32-bit and 64-bit builds. Cleaned up lint warnings regarding type mismatches. Deleted redundant casts.

+++ src/case_cmd.c
@@ -78,8 +78,8 @@ static void exec_case(struct cmd *cmd, bool lower)
 {
     assert(cmd != NULL);
 
-    int dot = (int)t.dot;
-    int Z   = (int)t.Z;
+    int dot = t.dot;
+    int Z   = t.Z;
     int m, n;
 
     if (cmd->h)                         // HFU/HFL?
