+++ src/fd_cmd.c
@@ -77,14 +77,14 @@ void exec_FD(struct cmd *cmd)
         s.search     = search_backward;
         s.count      = -cmd->n_arg;
         s.text_start = -1;              // Start at previous character
-        s.text_end = -(int)t.dot;
+        s.text_end   = -t.dot;
     }
     else
     {
         s.search     = search_forward;
         s.count      = cmd->n_arg;
         s.text_start = 0;               // Start at current character
-        s.text_end   = (int)(t.Z - t.dot);
+        s.text_end   = t.Z - t.dot;
     }
 
     if (search_loop(&s))
